Documentation + host instructions
