+++ .vscode/settings.json
@@ -1,3 +1,6 @@
 {
-    "cmake.sourceDirectory": "C:/Users/Hasan/OneDrive/Desktop/earth-rover-mini-OpenSource/Software/Linux"
+    "cmake.sourceDirectory": "C:/Users/Hasan/OneDrive/Desktop/earth-rover-mini-OpenSource/Software/Linux",
+    "files.associations": {
+        "rtthread.h": "c"
+    }
 }
