--- Software/Linux/src/hello.cpp
@@ -1,7 +1,0 @@
-#include <stdio.h>
-#include <stdlib.h>
-
-int main(int argc, char* argv[]) {
-  printf("Hello, World!\n");
-  return 0;
-}
